pyproject.toml
[project]
name = "mcp_creation"
version = "0.1.0"
description = "Creating a MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "fastapi[standard]>=0.135.3",
    "fastmcp>=3.2.3",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "mcp-server-time>=2026.1.26",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "uvicorn>=0.44.0",
]
